requirements.txt
discord
discord.py[voice]
psutil
requests
beautifulsoup4
python-dateutil
tortoise-orm
aiosqlite